Refactor RealmObjectSchema Clean up Proxy generation Comments addressed
