+++ realm/realm-annotations-processor/src/main/java/io/realm/processor/RealmProxyMediatorGenerator.java
@@ -38,9 +38,9 @@ import static io.realm.processor.Constants.REALM_PACKAGE_NAME;
 
 public class RealmProxyMediatorGenerator {
     private final String className;
-    private ProcessingEnvironment processingEnvironment;
-    private List<String> qualifiedModelClasses = new ArrayList<String>();
-    private List<String> qualifiedProxyClasses = new ArrayList<String>();
+    private final ProcessingEnvironment processingEnvironment;
+    private final List<String> qualifiedModelClasses = new ArrayList<String>();
+    private final List<String> qualifiedProxyClasses = new ArrayList<String>();
 
     public RealmProxyMediatorGenerator(ProcessingEnvironment processingEnvironment,
             String className, Set<ClassMetaData> classesToValidate) {
@@ -96,7 +96,6 @@ public class RealmProxyMediatorGenerator {
         writer.emitEmptyLine();
 
         emitFields(writer);
-        emitCreateTableMethod(writer);
         emitCreateRealmObjectSchema(writer);
         emitValidateTableMethod(writer);
         emitGetFieldNamesMethod(writer);
@@ -135,28 +134,10 @@ public class RealmProxyMediatorGenerator {
                 EnumSet.of(Modifier.PUBLIC),
                 "Class<? extends RealmModel>", "clazz", "RealmSchema", "realmSchema"
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return %s.createRealmObjectSchema(realmSchema)", qualifiedProxyClasses.get(i));
-            }
-        }, writer);
-        writer.endMethod();
-        writer.emitEmptyLine();
-    }
-
-    private void emitCreateTableMethod(JavaWriter writer) throws IOException {
-        writer.emitAnnotation("Override");
-        writer.beginMethod(
-                "Table",
-                "createTable",
-                EnumSet.of(Modifier.PUBLIC),
-                "Class<? extends RealmModel>", "clazz", "SharedRealm", "sharedRealm"
-        );
-        emitMediatorSwitch(new ProxySwitchStatement() {
-            @Override
-            public void emitStatement(int i, JavaWriter writer) throws IOException {
-                writer.emitStatement("return %s.initTable(sharedRealm)", qualifiedProxyClasses.get(i));
             }
         }, writer);
         writer.endMethod();
@@ -173,7 +154,7 @@ public class RealmProxyMediatorGenerator {
                 "SharedRealm", "sharedRealm",
                 "boolean", "allowExtraColumns"
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return %s.validateTable(sharedRealm, allowExtraColumns)",
@@ -192,7 +173,7 @@ public class RealmProxyMediatorGenerator {
                 EnumSet.of(Modifier.PUBLIC),
                 "Class<? extends RealmModel>", "clazz"
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return %s.getFieldNames()", qualifiedProxyClasses.get(i));
@@ -210,7 +191,7 @@ public class RealmProxyMediatorGenerator {
                 EnumSet.of(Modifier.PUBLIC),
                 "Class<? extends RealmModel>", "clazz"
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return %s.getTableName()", qualifiedProxyClasses.get(i));
@@ -236,7 +217,7 @@ public class RealmProxyMediatorGenerator {
         writer.emitStatement("final BaseRealm.RealmObjectContext objectContext = BaseRealm.objectContext.get()");
         writer.beginControlFlow("try")
                 .emitStatement("objectContext.set((BaseRealm) baseRealm, row, columnInfo, acceptDefaultValue, excludeFields)");
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return clazz.cast(new %s())", qualifiedProxyClasses.get(i));
@@ -269,7 +250,7 @@ public class RealmProxyMediatorGenerator {
         writer.emitSingleLineComment("generated by RealmProxy or the original type extending directly from RealmObject");
         writer.emitStatement("@SuppressWarnings(\"unchecked\") Class<E> clazz = (Class<E>) ((obj instanceof RealmObjectProxy) ? obj.getClass().getSuperclass() : obj.getClass())");
         writer.emitEmptyLine();
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return clazz.cast(%s.copyOrUpdate(realm, (%s) obj, update, cache))", qualifiedProxyClasses.get(i), qualifiedModelClasses.get(i));
@@ -412,7 +393,7 @@ public class RealmProxyMediatorGenerator {
                 Arrays.asList("Class<E>", "clazz", "Realm", "realm", "JSONObject", "json", "boolean", "update"),
                 Arrays.asList("JSONException")
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return clazz.cast(%s.createOrUpdateUsingJsonObject(realm, json, update))", qualifiedProxyClasses.get(i));
@@ -431,7 +412,7 @@ public class RealmProxyMediatorGenerator {
                 Arrays.asList("Class<E>", "clazz", "Realm", "realm", "JsonReader", "reader"),
                 Arrays.asList("java.io.IOException")
         );
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return clazz.cast(%s.createUsingJsonStream(realm, reader))", qualifiedProxyClasses.get(i));
@@ -453,7 +434,7 @@ public class RealmProxyMediatorGenerator {
         writer.emitSingleLineComment("generated by RealmProxy or the original type extending directly from RealmObject");
         writer.emitStatement("@SuppressWarnings(\"unchecked\") Class<E> clazz = (Class<E>) realmObject.getClass().getSuperclass()");
         writer.emitEmptyLine();
-        emitMediatorSwitch(new ProxySwitchStatement() {
+        emitMediatorShortCircuitSwitch(new ProxySwitchStatement() {
             @Override
             public void emitStatement(int i, JavaWriter writer) throws IOException {
                 writer.emitStatement("return clazz.cast(%s.createDetachedCopy((%s) realmObject, 0, maxDepth, cache))",
@@ -491,6 +472,26 @@ public class RealmProxyMediatorGenerator {
             writer.endControlFlow();
         }
     }
+
+    // Identical to the above, but eliminates the un-needed "else" clauses for, e.g., return statements
+    private void emitMediatorShortCircuitSwitch(ProxySwitchStatement statement, JavaWriter writer) throws IOException {
+        emitMediatorShortCircuitSwitch(statement, writer, true);
+    }
+
+    private void emitMediatorShortCircuitSwitch(ProxySwitchStatement statement, JavaWriter writer, boolean nullPointerCheck)
+            throws IOException {
+        if (nullPointerCheck) {
+            writer.emitStatement("checkClass(clazz)");
+            writer.emitEmptyLine();
+        }
+        for (int i = 0; i < qualifiedModelClasses.size(); i++) {
+            writer.beginControlFlow("if (clazz.equals(%s.class))", qualifiedModelClasses.get(i));
+            statement.emitStatement(i, writer);
+            writer.endControlFlow();
+        }
+        writer.emitStatement("throw getMissingProxyClassException(clazz)");
+    }
+
 
     private String getProxyClassName(String clazz) {
         return clazz + Constants.PROXY_SUFFIX;
