+++ realm/realm-library/src/main/cpp/CMakeLists.txt
@@ -37,7 +37,7 @@ set(classes_LIST
     io.realm.internal.LinkView io.realm.internal.Util io.realm.internal.UncheckedRow
     io.realm.internal.TableQuery io.realm.internal.SharedRealm io.realm.internal.TestUtil
     io.realm.log.LogLevel io.realm.log.RealmLog io.realm.Property io.realm.OsRealmSchema
-    io.realm.RealmObjectSchema io.realm.internal.Collection
+    io.realm.OsRealmObjectSchema io.realm.internal.Collection
     io.realm.internal.NativeObjectReference io.realm.internal.CollectionChangeSet
     io.realm.internal.OsObject
 )
